+++ src/test/converter2/issues/gh2704.ts
@@ -1,0 +1,4 @@
+export type Foo = Bar<123>;
+
+/** @ignore */
+export type Bar<T> = T extends 123 ? true : false;